FDG-6175 Move afg components
