+++ src/layouts/explainer/explainer-components/afg-components/afg-hero/afg-hero.module.scss
