+++ src/layouts/explainer/explainer-components/afg-components/afg-icon/afg-icon.jsx
@@ -1,6 +1,5 @@
 import React, {useState, useEffect} from 'react'
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faHandHoldingDollar } from "@fortawesome/free-solid-svg-icons";
 
 import {
     icon,
