+++ src/layouts/explainer/explainer-components/afg-components/afg-icon/afg-icon.module.scss
