+++ src/layouts/explainer/explainer-components/afg-components/afg-topic-section/afg-topic-section.jsx
@@ -1,11 +1,10 @@
 import React from 'react'
 import { Grid } from '@material-ui/core';
-import CustomLink from '../../../../components/links/custom-link/custom-link';
-import { ChartPlaceholder } from '../../../../layouts/explainer/explainer-helpers/national-deficit/national-deficit-helper';
+import { ChartPlaceholder } from
+    '../../../explainer-helpers/national-deficit/national-deficit-helper';
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faArrowRightLong} from "@fortawesome/free-solid-svg-icons";
 import * as styles from "./afg-topic-section.module.scss";
-import { style } from 'd3-selection';
 
 export default function AfgTopicSection({
     heading,
