+++ src/layouts/explainer/explainer-components/afg-components/afg-topic-section/afg-topic-section.module.scss.d.ts
