+++ src/layouts/explainer/explainer-components/afg-components/compare-section/compare-section.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../variables.module.scss';
+@import 'src/variables.module';
 
 .compareSection {
