+++ src/layouts/explainer/explainer-components/afg-components/compare-section/compare-section.spec.tsx
